+++ SemanticKernel.Assistants.sln
@@ -37,23 +37,32 @@ EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
+		Publish|Any CPU = Publish|Any CPU
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
 		{6D530B82-6217-4A67-B22D-E5ACFAE4A511}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{6D530B82-6217-4A67-B22D-E5ACFAE4A511}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6D530B82-6217-4A67-B22D-E5ACFAE4A511}.Publish|Any CPU.ActiveCfg = Publish|Any CPU
+		{6D530B82-6217-4A67-B22D-E5ACFAE4A511}.Publish|Any CPU.Build.0 = Publish|Any CPU
 		{6D530B82-6217-4A67-B22D-E5ACFAE4A511}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{6D530B82-6217-4A67-B22D-E5ACFAE4A511}.Release|Any CPU.Build.0 = Release|Any CPU
 		{03C21161-E835-4857-A81A-C1727140E920}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{03C21161-E835-4857-A81A-C1727140E920}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{03C21161-E835-4857-A81A-C1727140E920}.Publish|Any CPU.ActiveCfg = Release|Any CPU
+		{03C21161-E835-4857-A81A-C1727140E920}.Publish|Any CPU.Build.0 = Release|Any CPU
 		{03C21161-E835-4857-A81A-C1727140E920}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{03C21161-E835-4857-A81A-C1727140E920}.Release|Any CPU.Build.0 = Release|Any CPU
 		{BBC6C36F-DC43-4FD3-9706-ECA4738F8F57}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{BBC6C36F-DC43-4FD3-9706-ECA4738F8F57}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{BBC6C36F-DC43-4FD3-9706-ECA4738F8F57}.Publish|Any CPU.ActiveCfg = Release|Any CPU
+		{BBC6C36F-DC43-4FD3-9706-ECA4738F8F57}.Publish|Any CPU.Build.0 = Release|Any CPU
 		{BBC6C36F-DC43-4FD3-9706-ECA4738F8F57}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{BBC6C36F-DC43-4FD3-9706-ECA4738F8F57}.Release|Any CPU.Build.0 = Release|Any CPU
 		{CEAC598F-3763-467A-904D-C0C098DF7B9E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{CEAC598F-3763-467A-904D-C0C098DF7B9E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{CEAC598F-3763-467A-904D-C0C098DF7B9E}.Publish|Any CPU.ActiveCfg = Publish|Any CPU
+		{CEAC598F-3763-467A-904D-C0C098DF7B9E}.Publish|Any CPU.Build.0 = Publish|Any CPU
 		{CEAC598F-3763-467A-904D-C0C098DF7B9E}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{CEAC598F-3763-467A-904D-C0C098DF7B9E}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
