+++ src/Assistants/GroupThread/GroupThreadInstructions.handlebars
